Fixed some problems where the ruler view wouldn't show up correctly in 64bit land.

+++ src/NoodleLineNumberMarker.h
@@ -32,13 +32,13 @@
 
 @interface NoodleLineNumberMarker : NSRulerMarker
 {
-	unsigned		lineNumber;
+	NSUInteger		lineNumber;
 }
 
-- (id)initWithRulerView:(NSRulerView *)aRulerView lineNumber:(float)line image:(NSImage *)anImage imageOrigin:(NSPoint)imageOrigin;
+- (id)initWithRulerView:(NSRulerView *)aRulerView lineNumber:(CGFloat)line image:(NSImage *)anImage imageOrigin:(NSPoint)imageOrigin;
 
-- (void)setLineNumber:(unsigned)line;
-- (unsigned)lineNumber;
+- (void)setLineNumber:(NSUInteger)line;
+- (NSUInteger)lineNumber;
 
 
 @end
